requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
idna==2.6
odfpy==1.3.6
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2-binary==2.7.4
python-dateutil==2.7.3
pytz==2018.4
reportlab==3.4.0
requests==2.18.4
six==1.11.0
urllib3==1.22
